+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Removed `Command#infoOptionAction` and `infoOptionAction` option for `Command` constructor
 - Added `Command#messageRef()` and `Option#messageRef()` methods
 - Changed `Command#command()` to raise an exception when subcommand name already in use
+- Added `Command#clone()` method
 
 ## 2.0.1 (December 16, 2019)